requirements.txt
discord.py
audioop-lts
